HA_clients/requirements.txt
bme680>=2.0.0
paho-mqtt>=2.1.0